requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
requests>=2.31,<3.0
geopy>=2.4,<3.0
